3) add new hyperparameters (inference rate, KDE bandwidth, SNR, etc) 2) model fitting speed is dramatically improved and fitting is more efficient 1) update terms used to describe messages and targets 0) test reports, cleanups, and minor bugs.

+++ python/vrc/utils/count_spikes.py
@@ -23,7 +23,7 @@ def count_spikes(spike_trains: list,
 
   channels_cnt = len(spike_trains)
 
-  spike_counts = np.zeros((channels_cnt, duration * counting_freq + 1))
+  spike_counts = np.zeros((channels_cnt, int(duration * counting_freq + 1)))
 
   for i, spike_train in enumerate(spike_trains):
     for spike in spike_train:
